backend/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
Django==4.1.3
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
